# apps/backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/participant/{id}": {
      "get": {
        "operationId": "ParticipantController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            }
          }
        },
        "summary": "Get a participant by Id",
        "tags": [
          "participant"
        ]
      },
      "put": {
        "operationId": "ParticipantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParticipantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            }
          }
        },
        "summary": "Update the participant by Id",
        "tags": [
          "participant"
        ]
      },
      "delete": {
        "operationId": "ParticipantController_deleteParticipant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          }
        },
        "summary": "Remove a participant from the database",
        "tags": [
          "participant"
        ]
      }
    },
    "/participant": {
      "get": {
        "operationId": "ParticipantController_findByQuery",
        "parameters": [
          {
            "name": "lossesTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "150",
              "type": "string"
            }
          },
          {
            "name": "lossesFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "10",
              "type": "string"
            }
          },
          {
            "name": "winsTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "150",
              "type": "string"
            }
          },
          {
            "name": "winsFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "10",
              "type": "string"
            }
          },
          {
            "name": "eloTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1700",
              "type": "string"
            }
          },
          {
            "name": "eloFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "1200",
              "type": "string"
            }
          },
          {
            "name": "updatedTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2012.10.05",
              "type": "string"
            }
          },
          {
            "name": "updatedFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2012.10.05",
              "type": "string"
            }
          },
          {
            "name": "createdTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2012.10.05",
              "type": "string"
            }
          },
          {
            "name": "createdFrom",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2012.10.05",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "name",
                "type",
                "elo",
                "wins",
                "losses",
                "createdAt",
                "updatedAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "Individual",
                  "Team"
                ]
              }
            }
          },
          {
            "name": "itemsPerPage",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "currentPage",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantsResponse"
                }
              }
            }
          }
        },
        "summary": "Get all participants",
        "tags": [
          "participant"
        ]
      },
      "post": {
        "operationId": "ParticipantController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParticipantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            }
          }
        },
        "summary": "Create a new participant",
        "tags": [
          "participant"
        ]
      }
    },
    "/participant/{tournamentId}": {
      "post": {
        "operationId": "ParticipantController_addParticipantToTournament",
        "parameters": [
          {
            "name": "tournamentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParticipantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantResponse"
                }
              }
            }
          }
        },
        "summary": "Create a new participant for a tournament",
        "tags": [
          "participant"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tournament API",
    "description": "API for Tournament management",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "TournamentStatus": {
        "type": "string",
        "enum": [
          "Setup",
          "Started",
          "Over"
        ]
      },
      "ParticipantType": {
        "type": "string",
        "enum": [
          "Individual",
          "Team"
        ]
      },
      "PhaseType": {
        "type": "string",
        "enum": [
          "None",
          "GroupStage",
          "SingleElimination",
          "DoubleElimination",
          "RoundRobin",
          "Swiss"
        ]
      },
      "EliminationType": {
        "type": "string",
        "enum": [
          "Knockout",
          "Double"
        ]
      },
      "MatchType": {
        "type": "string",
        "enum": [
          "Group",
          "Knockout",
          "Bronze",
          "Final"
        ]
      },
      "ParticipantTournament": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "participantId": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournament": {
            "$ref": "#/components/schemas/Tournament"
          },
          "participant": {
            "$ref": "#/components/schemas/Participant"
          }
        },
        "required": [
          "id",
          "participantId",
          "tournamentId",
          "wins",
          "losses",
          "createdAt",
          "updatedAt"
        ]
      },
      "TournamentWinner": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "participantId": {
            "type": "integer",
            "format": "int32"
          },
          "place": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournament": {
            "$ref": "#/components/schemas/Tournament"
          },
          "participant": {
            "$ref": "#/components/schemas/Participant"
          }
        },
        "required": [
          "id",
          "tournamentId",
          "participantId",
          "place",
          "createdAt",
          "updatedAt"
        ]
      },
      "Participant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantType"
              }
            ]
          },
          "elo": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournaments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantTournament"
            }
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantGroup"
            }
          },
          "matchesAsP1": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          },
          "matchesAsP2": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          },
          "matchesWon": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          },
          "podiums": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentWinner"
            }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "elo",
          "wins",
          "losses",
          "createdAt",
          "updatedAt"
        ]
      },
      "ParticipantGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "participantId": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "group": {
            "$ref": "#/components/schemas/TournamentGroup"
          },
          "participant": {
            "$ref": "#/components/schemas/Participant"
          }
        },
        "required": [
          "id",
          "tournamentGroupId",
          "participantId",
          "wins",
          "losses",
          "points",
          "createdAt",
          "updatedAt"
        ]
      },
      "TournamentGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentPhaseId": {
            "type": "integer",
            "format": "int32"
          },
          "groupNumber": {
            "type": "integer",
            "format": "int32"
          },
          "isGroupMatchesEnded": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournamentPhase": {
            "$ref": "#/components/schemas/TournamentPhase"
          },
          "participantGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantGroup"
            }
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          }
        },
        "required": [
          "id",
          "tournamentPhaseId",
          "groupNumber",
          "isGroupMatchesEnded",
          "createdAt",
          "updatedAt"
        ]
      },
      "Match": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentPhaseId": {
            "type": "integer",
            "format": "int32"
          },
          "eliminationId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tournamentGroupId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "participant1Id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "participant2Id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "winnerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nextMatchId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "round": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "serialNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isOver": {
            "type": "boolean"
          },
          "matchType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MatchType"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournamentPhase": {
            "$ref": "#/components/schemas/TournamentPhase"
          },
          "elimination": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Elimination"
              }
            ]
          },
          "group": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TournamentGroup"
              }
            ]
          },
          "participant1": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Participant"
              }
            ]
          },
          "participant2": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Participant"
              }
            ]
          },
          "winner": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Participant"
              }
            ]
          },
          "nextMatch": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Match"
              }
            ]
          },
          "prevMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          }
        },
        "required": [
          "id",
          "tournamentPhaseId",
          "eliminationId",
          "tournamentGroupId",
          "participant1Id",
          "participant2Id",
          "winnerId",
          "nextMatchId",
          "round",
          "serialNumber",
          "isOver",
          "matchType",
          "createdAt",
          "updatedAt"
        ]
      },
      "Elimination": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentPhaseId": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EliminationType"
              }
            ]
          },
          "currentRound": {
            "type": "integer",
            "format": "int32"
          },
          "isOver": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tournamentPhase": {
            "$ref": "#/components/schemas/TournamentPhase"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          }
        },
        "required": [
          "id",
          "tournamentPhaseId",
          "type",
          "currentRound",
          "isOver",
          "createdAt",
          "updatedAt"
        ]
      },
      "TournamentPhase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "phaseType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PhaseType"
              }
            ]
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "elimination": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Elimination"
              }
            ]
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentGroup"
            }
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          },
          "tournament": {
            "$ref": "#/components/schemas/Tournament"
          }
        },
        "required": [
          "id",
          "tournamentId",
          "phaseType",
          "order",
          "isCompleted"
        ]
      },
      "Tournament": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TournamentStatus"
              }
            ]
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantType"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "phases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentPhase"
            }
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantTournament"
            }
          },
          "winners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentWinner"
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PaginationData": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "number",
            "example": 42
          },
          "totalItems": {
            "type": "number",
            "example": 42
          }
        },
        "required": [
          "totalPages",
          "totalItems"
        ]
      },
...
      "BaseResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ]
      },
      "ParticipantResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Participant"
          }
        },
        "required": [
          "ok"
        ]
      },
      "ParticipantsResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Participant"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationData"
          }
        },
        "required": [
          "ok"
        ]
      },
      "CreateParticipantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantType"
              }
            ]
          },
          "elo": {
            "type": "integer",
            "format": "int32",
            "default": 1500
          },
          "wins": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "losses": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "2025-10-11T20:55:39.335Z"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateParticipantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ParticipantType"
              }
            ]
          },
          "elo": {
            "type": "integer",
            "format": "int32",
            "default": 1500
          },
          "wins": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "losses": {
            "type": "integer",
            "format": "int32",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "default": "2025-10-11T20:55:39.355Z"
          }
        }
      },
...
    }
  }
}